Move test deployment out of `daphne_worker`
Defines a new crate, `daphne_worker_test`, that defines a local testing
deployment. The integration tests, which use this deployment, are moved
to the new crate.
