+++ daphne_worker/Cargo.toml
@@ -13,16 +13,10 @@ license = "BSD-3-Clause"
 [lib]
 crate-type = ["cdylib", "rlib"]
 
-[features]
-default = ["console_error_panic_hook"]
-test_e2e = []
-test_janus = []
-
 [dependencies]
 daphne = { path = "../daphne" }
 async-trait = "0.1.56"
 base64 = "0.13.0"
-cfg-if = "0.1.2"
 worker = "0.0.10"
 serde_json = "1.0.67"
 getrandom = { version = "0.2", features = ["js"] } # Required for prio
@@ -35,30 +29,3 @@ rand = "0.8.5"
 reqwest-wasm = { version = "0.11", features = ["json"] }
 thiserror = "1.0"
 url = { version = "2.2.2", features = ["serde"] }
-
-# The `console_error_panic_hook` crate provides better debugging of panics by
-# logging them with `console.error`. This is great for development, but requires
-# all the `std::fmt` and `std::panicking` infrastructure, so isn't great for
-# code size when deploying.
-console_error_panic_hook = { version = "0.1.1", optional = true }
-
-[dev-dependencies]
-reqwest = { version = "0.11.7", features = ["json"] }
-assert_matches = "1.5.0"
-lazy_static = "1.4.0"
-futures = "0.3.21"
-janus = { git = "https://github.com/divviup/janus", rev = "6bb0e992da50afc8a5eb602ad5ce559fe78cf2c4" }
-janus_server = { git = "https://github.com/divviup/janus", rev = "6bb0e992da50afc8a5eb602ad5ce559fe78cf2c4" }
-janus_test_util = { git = "https://github.com/divviup/janus", rev = "6bb0e992da50afc8a5eb602ad5ce559fe78cf2c4" }
-tokio = { version = "^1.9", features = ["full"] }
-
-# Dependencies for running janus_server.
-#
-# TODO Consider making the "janus_test" a workspace, as these depeendencies are
-# only required for Janus interop tests. Note that eventually we might be able
-# to just run Janus in a container,  at which point these dependencies would go
-# away.
-deadpool-postgres = "0.10.1"
-tracing = "0.1.34"
-testcontainers = "0.14.0"
-tokio-postgres = { version = "0.7.6", features = ["with-chrono-0_4", "with-serde_json-1", "with-uuid-1", "array-impls"] }
