+++ daphne_worker_test/Cargo.toml
@@ -1,0 +1,64 @@
+# SPDX-License-Identifier: BSD-3-Clause
+
+[package]
+name = "daphne-worker-test"
+version = "0.0.1"
+authors = [
+  "Christopher Patton <cpatton@cloudflare.com>",
+  "Armando Faz Hernandez <armfazh@cloudflare.com>",
+]
+edition = "2021"
+license = "BSD-3-Clause"
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[features]
+default = ["console_error_panic_hook"]
+test_e2e = []
+test_janus = []
+
+[dependencies]
+daphne_worker = { path = "../daphne_worker" }
+cfg-if = "0.1.2"
+worker = "0.0.10"
+serde_json = "1.0.67"
+
+# The `console_error_panic_hook` crate provides better debugging of panics by
+# logging them with `console.error`. This is great for development, but requires
+# all the `std::fmt` and `std::panicking` infrastructure, so isn't great for
+# code size when deploying.
+console_error_panic_hook = { version = "0.1.1", optional = true }
+
+[profile.release]
+# Tell `rustc` to optimize for small code size.
+opt-level = "s"
+
+[dev-dependencies]
+assert_matches = "1.5.0"
+base64 = "0.13.0"
+daphne = { path = "../daphne" }
+hex = { version = "0.4.3", features = ["serde"] }
+lazy_static = "1.4.0"
+futures = "0.3.21"
+janus = { git = "https://github.com/divviup/janus", rev = "6bb0e992da50afc8a5eb602ad5ce559fe78cf2c4" }
+janus_server = { git = "https://github.com/divviup/janus", rev = "6bb0e992da50afc8a5eb602ad5ce559fe78cf2c4" }
+janus_test_util = { git = "https://github.com/divviup/janus", rev = "6bb0e992da50afc8a5eb602ad5ce559fe78cf2c4" }
+prio = "0.8.0"
+rand = "0.8.5"
+reqwest = { version = "0.11.7", features = ["json"] }
+ring = "0.16.20"
+tokio = { version = "^1.9", features = ["full"] }
+url = { version = "2.2.2", features = ["serde"] }
+
+
+# Dependencies for running janus_server.
+#
+# TODO Consider making the "janus_test" a workspace, as these depeendencies are
+# only required for Janus interop tests. Note that eventually we might be able
+# to just run Janus in a container,  at which point these dependencies would go
+# away.
+deadpool-postgres = "0.10.1"
+tracing = "0.1.34"
+testcontainers = "0.14.0"
+tokio-postgres = { version = "0.7.6", features = ["with-chrono-0_4", "with-serde_json-1", "with-uuid-1", "array-impls"] }
