+++ daphne_worker_test/docker/runtests.Dockerfile
@@ -1,6 +1,6 @@
 FROM rust:1.61-bullseye
 
-WORKDIR /tmp/ppm
+WORKDIR /tmp/dap_test
 
 RUN apt-get update && \
     apt-get install -y \
@@ -10,9 +10,10 @@ RUN apt-get update && \
 RUN rustup component add clippy-preview
 
 COPY Cargo.toml Cargo.lock ./
+COPY daphne_worker_test ./daphne_worker_test
 COPY daphne_worker ./daphne_worker
 COPY daphne ./daphne
-COPY daphne_worker/tests/backend/test.sh /
+COPY daphne_worker_test/docker/test.sh /
 RUN chmod +x /test.sh
 
 RUN cargo build
