+++ daphne_worker_test/docker/test.sh
