+++ daphne_worker_test/src/utils.rs
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: BSD-3-Clause
 
 use cfg_if::cfg_if;
-use worker::*;
 
 cfg_if! {
     // https://github.com/rustwasm/console_error_panic_hook#readme
@@ -13,13 +12,4 @@ cfg_if! {
         #[inline]
         pub fn set_panic_hook() {}
     }
-}
-
-pub(crate) fn now() -> u64 {
-    Date::now().as_millis() / 1000
-}
-
-pub(crate) fn int_err<S: ToString>(s: S) -> Error {
-    console_error!("internal error: {}", s.to_string());
-    Error::RustError("internalError".to_string())
 }
